+++ CHANGELOG.md
@@ -2,11 +2,19 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-
-## 1.1.1
+## 2.0.0
 ##### Unreleased
-### Changed
-* Improved logging
+###
+* "Copy" constructors to all data model classes that use another instance to initialise
+
+### Changed 
+To create a better separation between the generic 4-Corner datamodel classes and the SMP client related code the project was split into two projects.
+This projects now only contains generic code not related to specific functionality. The SMP client code has been moved to a new project: 
+[BDXR SMP Client](https://github.com/holodeck-b2b/bdxr-smp-client)
+
+### Removed
+* The possibility to include multiple process identifiers in `org.holodeckb2b.bdxr.smp.datamodel.ProcessInfo` as this is not
+  possible in the SMP data model
 
 ## 1.1.0
 ##### 2019-03-06